Update exp system to work without the core changes, add another perk typo prerequisite fix
