+++ css/applications.less
@@ -1,4 +1,4 @@
-
+// main: main.less
 
 .application.sheet.actor.v2 .window-content .tab.active[data-tab=skills] .skill-component.ptr2e-modifications {
     .skill-list .skill {
