+++ css/chat.less
@@ -1,0 +1,52 @@
+// main: main.less
+
+.chat-message.message{
+
+    &.ptr2e-modifications\.experience {
+        .modifiers {
+            display: none;
+            flex-direction: column;
+            gap: 0.25em;
+            margin-top: 0.5em;
+            margin-bottom: 0.5em;
+
+            .circumstance-modifier {
+                display: flex;
+                justify-content: space-between;
+            }
+        }
+        .awarded-to {
+            display: none;
+            flex-direction: column;
+            gap: 0.25em;
+            margin-top: 0.5em;
+            margin-bottom: 0.5em;
+
+            .awarded-exp {
+                flex: 1;
+                width: 100%;
+                display: flex;
+                align-items: center;
+                gap: 1em;
+                --img-size: 32px;
+                .icon {
+                    height: var(--img-size);
+                    width: var(--img-size);
+                    border: none;
+                }
+                .text {
+                    width:100%;
+                    display: flex;
+                    flex-direction: column;
+                    align-items: center;
+
+                    .exp {
+                        width: 100%;
+                        display: flex;
+                        justify-content: space-between;
+                    }
+                }
+            }
+        }
+    }
+}
