+++ js/main.mjs
@@ -4,7 +4,7 @@ import * as perks from "./perks.mjs";
 import * as actions from "./actions.mjs";
 import * as clocks from "./clocks.mjs";
 import * as encounter from "./encounter-building.mjs";
-import * as exp from "./exp.mjs";
+import * as exp from "./experience/index.mjs";
 import * as tokenDisplay from "./token-display.mjs";
 import * as tokenDrop from "./token-drop.mjs";
 import * as skills from "./categorized-skill-system/index.mjs";
