+++ module.json
@@ -1,55 +1,60 @@
 {
-    "name": "ptr2e-modifications",
-    "title": "PTR 2e Modifications",
-    "description": "Modifications that haven't made it into the core system yet, and may never.",
-    "authors": [
-        {
-            "name": "RightHandOfVecna",
-            "email": "",
-            "url": ""
-        }
-    ],
-    "version": "0.1.9",
-    "compatibility": {
-        "minimum": "12.327",
-        "verified": "12.331"
-    },
-    "relationships": {
-        "systems": [
-            {
-                "id": "ptr2e",
-                "type": "system",
-                "manifest": "https://2e.ptr.wiki/foundry/download?repo=ptr2e&type=json",
-                "compatibility": {
-                    "verified": "0.10.0-alpha.3.2.0"
-                }
-            }
-        ],
-        "requires": [
-            {
-            "id": "lib-wrapper",
-            "type": "module",
-            "compatibility": {
-                "minimum": "1.0.0.0",
-                "verified": "1.12.6.0"
-            }
-            }
-        ]
-    },
-    "esmodules": [
-        "/js/main.mjs"
-    ],
-    "styles": [
-        "/css/applications.css"
-    ],
-    "languages": [
+  "name": "ptr2e-modifications",
+  "title": "PTR 2e Modifications",
+  "description": "Modifications that haven't made it into the core system yet, and may never.",
+  "authors": [
+    {
+      "name": "RightHandOfVecna",
+      "email": "",
+      "url": ""
+    }
+  ],
+  "version": "0.1.9",
+  "compatibility": {
+    "minimum": "12.327",
+    "verified": "12.331"
+  },
+  "relationships": {
+    "systems": [
       {
-        "lang": "en",
-        "name": "English",
-        "path": "lang/en.json"
+        "id": "ptr2e",
+        "type": "system",
+        "manifest": "https://2e.ptr.wiki/foundry/download?repo=ptr2e&type=json",
+        "compatibility": {
+          "verified": "0.10.0-alpha.4.1.0"
+        }
       }
     ],
-    "url": "https://github.com/righthandofvecna/ptr2e-modifications",
-    "manifest": "https://github.com/righthandofvecna/ptr2e-modifications/releases/latest/download/module.json",
-    "download": "https://github.com/righthandofvecna/ptr2e-modifications/releases/latest/download/module.zip"
+    "requires": [
+      {
+        "id": "lib-wrapper",
+        "type": "module",
+        "compatibility": {
+          "minimum": "1.0.0.0",
+          "verified": "1.12.6.0"
+        }
+      }
+    ]
+  },
+  "esmodules": [
+    "/js/main.mjs"
+  ],
+  "styles": [
+    "/css/main.css"
+  ],
+  "languages": [
+    {
+      "lang": "en",
+      "name": "English",
+      "path": "lang/en.json"
+    }
+  ],
+  "documentTypes": {
+    "ChatMessage": {
+      "experience": {}
+    }
+  },
+  "url": "https://github.com/righthandofvecna/ptr2e-modifications",
+  "manifest": "https://github.com/righthandofvecna/ptr2e-modifications/releases/latest/download/module.json",
+  "download": "https://github.com/righthandofvecna/ptr2e-modifications/releases/latest/download/module.zip"
 }
